Move old redirect files into old package
COSDK-690
