+++ action-core/api/action-core.api
@@ -43,8 +43,8 @@ public final class com/adyen/checkout/action/core/GenericActionConfiguration$Bui
 	public synthetic fun addAwaitActionConfiguration (Lcom/adyen/checkout/await/old/AwaitConfiguration;)Ljava/lang/Object;
 	public fun addQRCodeActionConfiguration (Lcom/adyen/checkout/qrcode/QRCodeConfiguration;)Lcom/adyen/checkout/action/core/GenericActionConfiguration$Builder;
 	public synthetic fun addQRCodeActionConfiguration (Lcom/adyen/checkout/qrcode/QRCodeConfiguration;)Ljava/lang/Object;
-	public fun addRedirectActionConfiguration (Lcom/adyen/checkout/redirect/RedirectConfiguration;)Lcom/adyen/checkout/action/core/GenericActionConfiguration$Builder;
-	public synthetic fun addRedirectActionConfiguration (Lcom/adyen/checkout/redirect/RedirectConfiguration;)Ljava/lang/Object;
+	public fun addRedirectActionConfiguration (Lcom/adyen/checkout/redirect/old/RedirectConfiguration;)Lcom/adyen/checkout/action/core/GenericActionConfiguration$Builder;
+	public synthetic fun addRedirectActionConfiguration (Lcom/adyen/checkout/redirect/old/RedirectConfiguration;)Ljava/lang/Object;
 	public fun addTwintActionConfiguration (Lcom/adyen/checkout/twint/action/TwintActionConfiguration;)Lcom/adyen/checkout/action/core/GenericActionConfiguration$Builder;
 	public synthetic fun addTwintActionConfiguration (Lcom/adyen/checkout/twint/action/TwintActionConfiguration;)Ljava/lang/Object;
 	public fun addVoucherActionConfiguration (Lcom/adyen/checkout/voucher/VoucherConfiguration;)Lcom/adyen/checkout/action/core/GenericActionConfiguration$Builder;
@@ -79,8 +79,8 @@ public abstract class com/adyen/checkout/action/core/internal/ActionHandlingPaym
 	public synthetic fun addAwaitActionConfiguration (Lcom/adyen/checkout/await/old/AwaitConfiguration;)Ljava/lang/Object;
 	public final fun addQRCodeActionConfiguration (Lcom/adyen/checkout/qrcode/QRCodeConfiguration;)Lcom/adyen/checkout/components/core/internal/BaseConfigurationBuilder;
 	public synthetic fun addQRCodeActionConfiguration (Lcom/adyen/checkout/qrcode/QRCodeConfiguration;)Ljava/lang/Object;
-	public final fun addRedirectActionConfiguration (Lcom/adyen/checkout/redirect/RedirectConfiguration;)Lcom/adyen/checkout/components/core/internal/BaseConfigurationBuilder;
-	public synthetic fun addRedirectActionConfiguration (Lcom/adyen/checkout/redirect/RedirectConfiguration;)Ljava/lang/Object;
+	public final fun addRedirectActionConfiguration (Lcom/adyen/checkout/redirect/old/RedirectConfiguration;)Lcom/adyen/checkout/components/core/internal/BaseConfigurationBuilder;
+	public synthetic fun addRedirectActionConfiguration (Lcom/adyen/checkout/redirect/old/RedirectConfiguration;)Ljava/lang/Object;
 	public final fun addTwintActionConfiguration (Lcom/adyen/checkout/twint/action/TwintActionConfiguration;)Lcom/adyen/checkout/components/core/internal/BaseConfigurationBuilder;
 	public synthetic fun addTwintActionConfiguration (Lcom/adyen/checkout/twint/action/TwintActionConfiguration;)Ljava/lang/Object;
 	public final fun addVoucherActionConfiguration (Lcom/adyen/checkout/voucher/VoucherConfiguration;)Lcom/adyen/checkout/components/core/internal/BaseConfigurationBuilder;
