+++ action-core/src/main/java/com/adyen/checkout/action/core/GenericActionConfiguration.kt
@@ -10,7 +10,6 @@ package com.adyen.checkout.action.core
 
 import android.content.Context
 import androidx.annotation.RestrictTo
-import com.adyen.checkout.action.core.GenericActionConfiguration.Builder
 import com.adyen.checkout.action.core.internal.ActionHandlingConfigurationBuilder
 import com.adyen.checkout.adyen3ds2.Adyen3DS2Configuration
 import com.adyen.checkout.await.old.AwaitConfiguration
@@ -22,7 +21,7 @@ import com.adyen.checkout.components.core.internal.BaseConfigurationBuilder
 import com.adyen.checkout.components.core.internal.Configuration
 import com.adyen.checkout.core.old.Environment
 import com.adyen.checkout.qrcode.QRCodeConfiguration
-import com.adyen.checkout.redirect.RedirectConfiguration
+import com.adyen.checkout.redirect.old.RedirectConfiguration
 import com.adyen.checkout.twint.action.TwintActionConfiguration
 import com.adyen.checkout.voucher.VoucherConfiguration
 import com.adyen.checkout.wechatpay.WeChatPayActionConfiguration
