+++ action-core/src/main/java/com/adyen/checkout/action/core/internal/provider/ActionComponentExtensions.kt
@@ -14,7 +14,7 @@ import com.adyen.checkout.components.core.action.Action
 import com.adyen.checkout.components.core.internal.provider.ActionComponentProvider
 import com.adyen.checkout.core.old.internal.util.runCompileOnly
 import com.adyen.checkout.qrcode.QRCodeComponent
-import com.adyen.checkout.redirect.RedirectComponent
+import com.adyen.checkout.redirect.old.RedirectComponent
 import com.adyen.checkout.twint.action.TwintActionComponent
 import com.adyen.checkout.voucher.VoucherComponent
 import com.adyen.checkout.wechatpay.WeChatPayActionComponent
