+++ example-app/src/main/java/com/adyen/checkout/example/service/ExampleAdvancedDropInService.kt
@@ -38,7 +38,7 @@ import com.adyen.checkout.example.extensions.toStringPretty
 import com.adyen.checkout.example.repositories.AddressLookupCompletionResult
 import com.adyen.checkout.example.repositories.AddressLookupRepository
 import com.adyen.checkout.example.repositories.PaymentsRepository
-import com.adyen.checkout.redirect.RedirectComponent
+import com.adyen.checkout.redirect.old.RedirectComponent
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
