+++ example-app/src/main/java/com/adyen/checkout/example/ui/giftcard/GiftCardActivity.kt
@@ -23,7 +23,7 @@ import com.adyen.checkout.example.databinding.ActivityGiftCardBinding
 import com.adyen.checkout.example.extensions.applyInsetsToRootLayout
 import com.adyen.checkout.example.ui.configuration.CheckoutConfigurationProvider
 import com.adyen.checkout.giftcard.GiftCardComponent
-import com.adyen.checkout.redirect.RedirectComponent
+import com.adyen.checkout.redirect.old.RedirectComponent
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
 import javax.inject.Inject
