+++ example-app/src/main/java/com/adyen/checkout/example/ui/instant/InstantFragment.kt
@@ -26,7 +26,7 @@ import com.adyen.checkout.components.core.action.Action
 import com.adyen.checkout.example.databinding.FragmentInstantBinding
 import com.adyen.checkout.example.ui.configuration.CheckoutConfigurationProvider
 import com.adyen.checkout.instant.InstantPaymentComponent
-import com.adyen.checkout.redirect.RedirectComponent
+import com.adyen.checkout.redirect.old.RedirectComponent
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
