+++ redirect/src/test/java/com/adyen/checkout/redirect/internal/data/api/NativeRedirectServiceTest.kt
@@ -1,7 +1,8 @@
 package com.adyen.checkout.redirect.internal.data.api
 
 import com.adyen.checkout.core.old.internal.data.api.HttpClient
-import com.adyen.checkout.redirect.internal.data.model.NativeRedirectRequest
+import com.adyen.checkout.redirect.old.internal.data.api.NativeRedirectService
+import com.adyen.checkout.redirect.old.internal.data.model.NativeRedirectRequest
 import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
